requirements.txt
# Dependencies for running the benchmark harness and agent
python-dotenv
google-generativeai